requirements.txt
Django==1.9.6
djangorestframework==3.3.3
Pygments==2.1.3
django-filter==0.13
numpy==1.11.0
simpy==3.0.8